SCRUM-91: Create UI for reset password page
Created a UI page for users to reset their password in case they forgot
their current password.
Implemented an error message UX for the text field in the case where the
provided email isn't in the database.
The code to check that an email is valid needs to be implemented.