SCRUM-90: Create UI for forgot username page
Created a UI page for users to get sent an email that contains the
username associated with their email in the case they forgot their
current username.
Implemented an error message UX for the text field in the case where the
provided email isn't in the database.
The code to check that an email is valid needs to be implemented.