SCRUM-86: Add missing UI buttons to existing pages
Created AddButton.jsx in the components folder. This is a template for a plus surrounded by a circle icon button.
Created SettingsButton.jsx in the components folder. This is a template
for a sprocket icon button.
Added the missing settings button and add button to HomePage.jsx.
Added the missing settings button and FPV/Control button to
ActivityLog.jsx.
Added the missing back button to ControlPage.jsx.