SCRUM-87: Adding button navigation paths
Added appropriate navigation paths for buttons as outlined by Figma.
The sign up button in NewAccount.jsx currently navigates to HomePage.jsx as it is assumed that we want users to be automatically logged in after signing up. If we want users to log in after signing up, this button should instead navigate to LoginPage.jsx,
Navigation path for the settings button in HomePage.jsx needs to be
added.