SCRUM-138: JointStates
Added some code that publishes a JointState message once.
Ideally, this would be linked to a timer to continually send JointState messages to more accurately detect the robot's position and account for things like wheel slippage.
JointState is necessary for an Odometry node which will likely need to be implemented.
The current build is very specific to 1 configuration of a robot.
To generalize, implementing ros2_control and SystemInterface could be considered. The user would then provide a URDF and YAML file for each robot.